ref: 화살표 함수로 수정 및 파라미터 수정

+++ src/controllers/product.controller.ts
@@ -1,7 +1,7 @@
 import type { NextFunction, Request, Response } from 'express';
 
 import { STATUS_CODE } from '../constants/constant.js';
-import type { CreateProductInput, GetProductListQuery, UpdateProductInput } from '../dtos/product.dto.js';
+import type { GetProductListQuery } from '../dtos/product.dto.js';
 import { ProductResponse } from '../serializes/product.serialize.js';
 import productService from '../services/product.service.js';
 import { HttpException } from '../utils/http-exception.js';
@@ -14,8 +14,7 @@ class ProductController {
         throw HttpException.unauthorized();
       }
 
-      const input: CreateProductInput = req.body;
-      const product = await productService.create(req.user.id, input);
+      const product = await productService.create({ userId: req.user.id, data: req.body });
       res.status(STATUS_CODE.CREATED).json(ProductResponse.detail(product));
     } catch (err) {
       next(err);
@@ -25,10 +24,11 @@ class ProductController {
   // 상품 목록 조회
   getList = async (req: Request, res: Response, next: NextFunction) => {
     try {
-      const filters: GetProductListQuery = req.query as unknown as GetProductListQuery;
-      const userId = req.user?.id;
-      const { list, totalCount } = await productService.getList(filters, userId);
-      res.status(STATUS_CODE.OK).json(ProductResponse.list(list, totalCount));
+      const result = await productService.getList({
+        filters: req.validated?.query as GetProductListQuery,
+        userId: req.user?.id,
+      });
+      res.status(STATUS_CODE.OK).json(ProductResponse.list(result.list, result.totalCount));
     } catch (err) {
       next(err);
     }
@@ -37,11 +37,10 @@ class ProductController {
   // 상품 상세 조회
   getById = async (req: Request, res: Response, next: NextFunction) => {
     try {
-      const { productId } = req.params;
-      if (!productId) {
+      if (!req.params.productId) {
         throw new HttpException({ status: STATUS_CODE.BAD_REQUEST, message: '상품 ID가 필요합니다' });
       }
-      const product = await productService.getById(productId);
+      const product = await productService.getById(req.params.productId);
       res.status(STATUS_CODE.OK).json(ProductResponse.detail(product));
     } catch (err) {
       next(err);
@@ -55,12 +54,15 @@ class ProductController {
         throw HttpException.unauthorized();
       }
 
-      const { productId } = req.params;
-      if (!productId) {
+      if (!req.params.productId) {
         throw new HttpException({ status: STATUS_CODE.BAD_REQUEST, message: '상품 ID가 필요합니다' });
       }
-      const input: UpdateProductInput = req.body;
-      const product = await productService.update(req.user.id, productId, input);
+
+      const product = await productService.update({
+        userId: req.user.id,
+        productId: req.params.productId,
+        ...req.body,
+      });
       res.status(STATUS_CODE.OK).json(ProductResponse.detail(product));
     } catch (err) {
       next(err);
@@ -74,11 +76,10 @@ class ProductController {
         throw HttpException.unauthorized();
       }
 
-      const { productId } = req.params;
-      if (!productId) {
+      if (!req.params.productId) {
         throw new HttpException({ status: STATUS_CODE.BAD_REQUEST, message: '상품 ID가 필요합니다' });
       }
-      await productService.delete(req.user.id, productId);
+      await productService.delete({ userId: req.user.id, productId: req.params.productId });
       res.status(STATUS_CODE.NO_CONTENT).send();
     } catch (err) {
       next(err);
